+++ src/test/ExSln2/LedBlinker/test_stuff/FuncPtrEx2.cs
@@ -5,8 +5,27 @@ namespace ts; // test stuff
 public class FuncPtrEx2 : FinObj
 {
     public delegate i32 FuncPtr(i32 a, i32 b);
-
     public FuncPtr func = add;
+
+    public class Bike : FinObj
+    {
+        public i32 speed;
+    }
+
+    /// <summary>
+    /// This is a delegate that takes a Bike and returns an i32.
+    /// </summary>
+    public delegate i32 BikeSpeedFunc(Bike bike);
+
+    /// <summary>
+    /// This is a delegate that takes a Bike and returns a Bike. Maybe for method chaining, but really just for testing :)
+    /// </summary>
+    public delegate Bike BikeBikeFunc(Bike bike);
+
+    /// <summary>
+    /// This is a delegate that takes a Bike and returns a Bike. Maybe for method chaining, but really just for testing :)
+    /// </summary>
+    public delegate c_array<Bike> BikeArrayFunc(c_array<Bike> bikes, u8 length);
 
     public static i32 add(i32 a, i32 b)
     {