+++ src/test/ExSln2/c99/gen/hal_LedArray.h
@@ -4,9 +4,9 @@
 
 #pragma once
 
+#include "hal_Led.h"
 #include <stdint.h>
 #include "hal_LedArray_LedInfo.h"
-#include "hal_Led.h"