+++ CHANGELOG.md
@@ -22,6 +22,14 @@ You can easily find changes by searching for `"### Change"` in this markdown fil
 # Releases
 Test/interim releases are not documented here.
 
+
+## [0.???]
+### Fixed
+- fix delegate and pointer type parameters
+    - https://github.com/fin-language/fin/issues/86
+
+---
+
 ## [0.6.1]
 ### Added
 - Add `c_array_mem<T>`